+++ cmd/skywire-cli/commands/config/gen.go
@@ -104,10 +104,10 @@ var genConfigCmd = &cobra.Command{
 			logger.WithError(err).Fatal("Failed to create config.")
 		}
 
-		// Use local servers
+		// Use dmsg urls for services and add dmsg-servers
 		if dmsgHTTP {
 			var dmsgHTTPServersList visorconfig.DmsgHTTPServers
-			serversListJSON, err := ioutil.ReadFile("localServers.json")
+			serversListJSON, err := ioutil.ReadFile("dmsghttp-config.json")
 			if err != nil {
 				logger.WithError(err).Fatal("Failed to read servers.json file.")
 			}