+++ pkg/servicedisc/client.go
@@ -48,11 +48,11 @@ type HTTPClient struct {
 	conf    Config
 	entry   Service
 	entryMx sync.Mutex // only used if RegisterEntry && DeleteEntry functions are used.
-	client  http.Client
+	client  *http.Client
 }
 
 // NewClient creates a new HTTPClient.
-func NewClient(log logrus.FieldLogger, mLog *logging.MasterLogger, conf Config) *HTTPClient {
+func NewClient(log logrus.FieldLogger, mLog *logging.MasterLogger, conf Config, client *http.Client) *HTTPClient {
 	return &HTTPClient{
 		log:  log,
 		mLog: mLog,
@@ -62,7 +62,7 @@ func NewClient(log logrus.FieldLogger, mLog *logging.MasterLogger, conf Config) 
 			Type:    conf.Type,
 			Version: buildinfo.Version(),
 		},
-		client: http.Client{},
+		client: client,
 	}
 }
 
@@ -99,7 +99,7 @@ func (c *HTTPClient) Auth(ctx context.Context) (*httpauth.Client, error) {
 		return auth, nil
 	}
 
-	auth, err := httpauth.NewClient(ctx, c.conf.DiscAddr, c.conf.PK, c.conf.SK, c.mLog)
+	auth, err := httpauth.NewClient(ctx, c.conf.DiscAddr, c.conf.PK, c.conf.SK, c.client, c.mLog)
 	if err != nil {
 		return nil, err
 	}