+++ pkg/setup/node.go
@@ -34,7 +34,7 @@ func NewNode(conf *Config) (*Node, error) {
 	}
 
 	// Connect to dmsg network.
-	dmsgDisc := disc.NewHTTP(conf.Dmsg.Discovery, http.Client{}, nil)
+	dmsgDisc := disc.NewHTTP(conf.Dmsg.Discovery, &http.Client{}, nil)
 	dmsgConf := &dmsg.Config{MinSessions: conf.Dmsg.SessionsCount}
 	dmsgC := dmsg.NewClient(conf.PK, conf.SK, dmsgDisc, dmsgConf)
 	go dmsgC.Serve(context.Background())