+++ pkg/transport/network/addrresolver/client_test.go
@@ -46,7 +46,7 @@ func TestClientAuth(t *testing.T) {
 
 	defer srv.Close()
 	log := logging.MustGetLogger("test_client_auth")
-	apiClient, err := NewHTTP(srv.URL, testPubKey, testSecKey, log, masterLogger)
+	apiClient, err := NewHTTP(srv.URL, testPubKey, testSecKey, &http.Client{}, log, masterLogger)
 	require.NoError(t, err)
 
 	c := apiClient.(*httpClient)
@@ -75,7 +75,7 @@ func TestBind(t *testing.T) {
 
 	defer srv.Close()
 	log := logging.MustGetLogger("test_bind")
-	c, err := NewHTTP(srv.URL, testPubKey, testSecKey, log, masterLogger)
+	c, err := NewHTTP(srv.URL, testPubKey, testSecKey, &http.Client{}, log, masterLogger)
 	require.NoError(t, err)
 
 	err = c.BindSTCPR(context.TODO(), "1234")